+++ src/main/java/com/napier/sem/App.java
@@ -1,64 +1,12 @@
 package com.napier.sem;
 
-import java.sql.*;
-
 public class App
 {
-    private Connection con = null;
-
-    public void connect()
-    {
-        try
-        {
-            Class.forName("com.mysql.cj.jdbc.Driver");
-        }
-        catch (ClassNotFoundException e)
-        {
-            System.out.println("Could not load SQL driver");
-            System.exit(-1);
-        }
-
-        int retries = 10;
-        for (int i = 0; i < retries; ++i)
-        {
-            System.out.println("Connecting to database...");
-            try
-            {
-                Thread.sleep(5000);
-                con = DriverManager.getConnection("jdbc:mysql://db:3306/world?useSSL=false", "root", "example");
-                System.out.println("Successfully connected");
-                break;
-            }
-            catch (SQLException sqle)
-            {
-                System.out.println("Failed to connect to database attempt " + Integer.toString(i));
-                System.out.println(sqle.getMessage());
-            }
-            catch (InterruptedException ie)
-            {
-                System.out.println("Thread interrupted? Should not happen.");
-            }
-        }
-    }
-    public void disconnect()
-    {
-        if (con != null)
-        {
-            try
-            {
-                con.close();
-            }
-            catch (Exception e)
-            {
-                System.out.println("Error closing connection to database");
-            }
-        }
-    }
 
     public static void main(String[] args)
     {
-        App a = new App();
-        a.connect();
-        a.disconnect();
+        // Create our new Application and run it
+        final Application app = new Application();
+        app.run();
     }
 }
